logic building

+++ src/utils/cloudinary.js
@@ -1,16 +1,17 @@
 import { v2 as cloudinary } from 'cloudinary';
 import fs from "fs";  
 
-(async function(localFilePath) {
 
-    // Configuration
-    cloudinary.config({ 
+const uploadOnCloudinary = async function(localFilePath) {
+      // Configuration
+        cloudinary.config({ 
         cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
         api_key: process.env.CLOUDINARY_API_KEY, 
         api_secret: process.env.CLOUDINARY_API_SECRET
      // Click 'View API Keys' above to copy your API secret
     });
     
+    if(!localFilePath) return null;
     // Upload an image
      const uploadResult = await cloudinary.uploader
        .upload(localFilePath, {
@@ -20,26 +21,30 @@ import fs from "fs";
        )
        .catch((error) => {
            console.log(error);
-           fs.unlinkSync(localFilePath);
+           if(localFilePath)fs.unlinkSync(localFilePath);
+           return null;
        });
     
-    console.log(uploadResult);
+    //console.log(uploadResult);
+    return uploadResult;
     
     // Optimize delivery by resizing and applying auto-format and auto-quality
-    const optimizeUrl = cloudinary.url('shoes', {
-        fetch_format: 'auto',
-        quality: 'auto'
-    });
+    // const optimizeUrl = cloudinary.url(uploadResult.public_id, {
+    //     fetch_format: 'auto',
+    //     quality: 'auto'
+    // });
     
-    console.log(optimizeUrl);
+    // console.log(optimizeUrl);
     
     // Transform the image: auto-crop to square aspect_ratio
-    const autoCropUrl = cloudinary.url('shoes', {
-        crop: 'auto',
-        gravity: 'auto',
-        width: 500,
-        height: 500,
-    });
+    // const autoCropUrl = cloudinary.url('shoes', {
+    //     crop: 'auto',
+    //     gravity: 'auto',
+    //     width: 500,
+    //     height: 500,
+    // });
     
-    console.log(autoCropUrl);    
-})();
+    // console.log(autoCropUrl);    
+};
+
+export default uploadOnCloudinary;
